pyproject.toml
[project]
name = "wargame"
version = "0.1.0"
description = "Generative AI geopolitical wargame — hybrid programmatic state engine + LLM adjudication"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "jupyter>=1.0",
    "matplotlib>=3.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["wargame*"]
